Upload images to cloudinary

+++ server/src/controllers/blogControllers.js
@@ -1,8 +1,9 @@
 const Blog = require('../models/Blog');
-const Photo = require('../models/Photo');
 const Post = require('../models/Post');
-const createPhotoLink = require('../utils/createPhotoLink');
 const { ErrorHandler } = require('../utils/error');
+const { deleteImageFromCloudinary } = require('../utils/cloudinary');
+const { uploader } = require('../services/cloudinary');
+const { dataUri } = require('../middleware/multer');
 
 exports.createBlog = async (req, res, next) => {
   const { name } = req.body;
@@ -10,7 +11,7 @@ exports.createBlog = async (req, res, next) => {
   try {
     const blog = await Blog.findOne({ name });
     if (blog) {
-      throw new ErrorHandler(400, `blog '${name}' already exists`);
+      throw new ErrorHandler(422, 'blog name already in use');
     }
 
     const blogData = { user: req.user._id, name, bgImg: {} };
@@ -19,86 +20,104 @@ exports.createBlog = async (req, res, next) => {
       blogData.description = req.body.description;
     }
 
-    if (typeof req.body.bgImgUrl !== 'undefined') {
-      blogData.bgImg.photoURL = req.body.bgImgUrl;
-      blogData.bgImg.photoID = null;
-    }
+    if (req.file || req.body.bgImgUrl) {
+      let image;
+      if (req.file) {
+        const file = dataUri(req).content;
+        image = file;
+      } else {
+        image = req.body.bgImgUrl;
+      }
 
-    if (typeof req.body.imgAttribution !== 'undefined') {
-      blogData.bgImg.imgAttribution = req.body.imgAttribution;
-    }
-
-    const file = req.file && req.file.buffer;
-    if (file && !blogData.bgImg.photoURL) {
-      const photo = await Photo.create({
-        photo: req.file.buffer,
+      const result = await uploader.upload(image, {
+        upload_preset: 'bloggingplatform',
       });
 
-      blogData.bgImg.photoID = photo.id;
-      blogData.bgImg.photoURL = createPhotoLink(photo.id);
+      blogData.bgImg.image_url = result.secure_url;
+      blogData.bgImg.large_image_url = result.eager[0].secure_url;
+    }
+
+    if (
+      blogData.bgImg.image_url &&
+      typeof req.body.imgAttribution !== 'undefined'
+    ) {
+      blogData.bgImg.img_attribution = req.body.imgAttribution;
     }
 
     const newBlog = await Blog.create({ ...blogData });
 
-    res.status(201).json({ blog: newBlog });
+    return res.status(201).json({ blog: newBlog });
   } catch (err) {
     next(err);
   }
 };
 
 exports.updateBlog = async (req, res, next) => {
-  try {
-    const blog = await Blog.findOne({
-      _id: req.params.blogId,
-      user: req.user._id,
-    });
+  const { blogId } = req.params;
 
+  try {
+    const blog = await Blog.findById(blogId);
     if (!blog) {
       throw new ErrorHandler(404, 'Blog Not Found');
     }
 
-    const blogData = { name: req.body.name, bgImg: {} };
+    if (!blog.user.equals(req.user._id)) {
+      throw new ErrorHandler(403, 'You are not allowed to update the blog');
+    }
+
+    if (blog.name.toLowerCase() !== req.body.name.toLowerCase()) {
+      const doc = await Blog.findOne({ name: req.body.name });
+      if (doc) {
+        throw new ErrorHandler(422, 'blog name aready in use');
+      }
+    }
+
+    const blogData = { name: req.body.name, bgImg: { ...blog.bgImg } };
 
     if (typeof req.body.description !== 'undefined') {
       blogData.description = req.body.description;
     }
 
-    if (typeof req.body.bgImgUrl !== 'undefined') {
-      blogData.bgImg.photoURL = req.body.bgImgUrl;
-      blogData.bgImg.photoID = null;
-    }
+    if (req.file || req.body.bgImgUrl) {
+      let image;
+      if (req.file) {
+        const file = dataUri(req).content;
+        image = file;
+      } else {
+        image = req.body.bgImgUrl;
+      }
 
-    if (typeof req.body.imgAttribution !== 'undefined') {
-      blogData.bgImg.imgAttribution = req.body.imgAttribution;
-    }
-
-    const newPhoto = req.file && req.file.buffer;
-
-    if (
-      blog.bgImg &&
-      blog.bgImg.photoID &&
-      (blogData.bgImg.photoURL || newPhoto)
-    ) {
-      // delete old photo
-      await Photo.findByIdAndDelete(blog.bgImg.photoID);
-    }
-
-    if (newPhoto) {
-      const photo = await Photo.create({
-        photo: newPhoto,
+      const result = await uploader.upload(image, {
+        upload_preset: 'bloggingplatform',
       });
 
-      blogData.bgImg.photoID = photo.id;
-      blogData.bgImg.photoURL = createPhotoLink(photo.id);
+      blogData.bgImg.image_url = result.secure_url;
+      blogData.bgImg.large_image_url = result.eager[0].secure_url;
     }
 
-    const updatedBlog = await Blog.findByIdAndUpdate(
-      req.params.blogId,
-      blogData,
-      { new: true }
-    );
+    if (
+      blogData.bgImg.image_url &&
+      typeof req.body.imgAttribution !== 'undefined'
+    ) {
+      blogData.bgImg.img_attribution = req.body.imgAttribution;
+    }
 
-    res.status(200).json({ blog: updatedBlog });
+    const updatedBlog = await Blog.findByIdAndUpdate(blogId, blogData, {
+      new: true,
+    });
+
+    // Delete old image
+    if (
+      blog.bgImg &&
+      blog.bgImg.image_url &&
+      updatedBlog.bgImg &&
+      updatedBlog.bgImg.image_url &&
+      blog.bgImg.image_url !== updatedBlog.bgImg.image_url
+    ) {
+      await deleteImageFromCloudinary(blog.bgImg.image_url, 'bloggingplatform');
+    }
+
+    return res.status(200).json({ blog: updatedBlog });
   } catch (err) {
     next(err);
   }
@@ -116,7 +135,7 @@ exports.getBlogById = async (req, res, next) => {
       throw new ErrorHandler(404, 'Blog Not Found');
     }
 
-    res.json({ blog });
+    return res.json({ blog });
   } catch (err) {
     next(err);
   }
@@ -134,7 +153,8 @@ exports.getBlogBySlugName = async (req, res, next) => {
     if (!blog) {
       throw new ErrorHandler(404, 'Blog Not Found');
     }
-    res.json({ blog });
+
+    return res.json({ blog });
   } catch (err) {
     next(err);
   }
@@ -147,7 +167,8 @@ exports.getAllBlogs = async (req, res, next) => {
       'bio',
       'avatar',
     ]);
-    res.json({ blogs });
+
+    return res.json({ blogs });
   } catch (err) {
     next(err);
   }
@@ -160,7 +181,8 @@ exports.getAuthUserBlogs = async (req, res, next) => {
       'bio',
       'avatar',
     ]);
-    res.json({ blogs });
+
+    return res.json({ blogs });
   } catch (err) {
     next(err);
   }
@@ -172,36 +194,45 @@ exports.getUserBlogs = async (req, res, next) => {
       'user',
       ['name', 'bio', 'avatar']
     );
-    res.json({ blogs });
+
+    return res.json({ blogs });
   } catch (err) {
     next(err);
   }
 };
 
 exports.deleteBlog = async (req, res, next) => {
+  const { blogId } = req.params;
+
   try {
-    const blog = await Blog.findOne({
-      _id: req.params.blogId,
-      user: req.user._id,
-    });
+    const blog = await Blog.findById(blogId);
 
     if (!blog) {
       throw new ErrorHandler(404, 'Blog not found');
     }
 
-    await Blog.deleteOne({ _id: req.params.blogId });
-
-    if (blog.bgImg && blog.bgImg.photoID) {
-      await Photo.deleteOne({ _id: blog.bgImg.photoID });
+    if (!blog.user.equals(req.user._id)) {
+      throw new ErrorHandler(403, 'You are not authorized to delete a blog');
     }
+
+    await Blog.deleteOne({ _id: blogId });
+
+    if (blog.bgImg && blog.bgImg.image_url) {
+      await deleteImageFromCloudinary(blog.bgImg.image_url, 'bloggingplatform');
+    }
+
     const posts = await Post.find({ blog: blog._id }).exec();
     posts.forEach(async (post) => {
-      if (post.photo) {
-        await Photo.deleteOne({ _id: post.photo });
-      }
       await Post.deleteOne({ _id: post._id });
+      if (post.bgImg && post.bgImg.image_url) {
+        await deleteImageFromCloudinary(
+          post.bgImg.image_url,
+          'bloggingplatform'
+        );
+      }
     });
-    res.status(201).json({ message: 'Blog deleted', blog });
+
+    return res.status(200).json({ message: 'Blog deleted', blog });
   } catch (err) {
     next(err);
   }
